adds index of longest run problem
Write a function that finds the zero-based index of the longest run in
a string. A run is a consecutive sequence of the same character. If
there is more than one run with the same length, return the index of
the first one.

For example, Run::indexOfLongestRun('abbcccddddcccbba') should return 6
as the longest run is dddd and it first appears on index 6.

https://www.testdome.com/Questions/PHP/Run/2983